consolidated messaging helpers into the utils/helpers file
